+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 1.4.0
+* Implement `Subscribe/Unsubscribe Characteristic`.
+
 ## 1.1.0
-* Test for Read/Write Characteristic 
+* Implement `Read/Write Characteristic`.
 
 ## 1.0.5
 * Change `rxdart` version to `0.27.7`.